Minor cleanup and refactoring
